+++ orbs/payloads.go
@@ -9,6 +9,9 @@ package orbs
 import (
 	"github.com/orbs-network/orbs-client-sdk-go/codec"
 	"github.com/orbs-network/orbs-client-sdk-go/crypto/encoding"
+	"io/ioutil"
+	"path"
+	"strings"
 	"time"
 )
 
@@ -89,4 +92,26 @@ func (c *OrbsClient) createGetBlockPayload(blockHeight uint64) (payload []byte, 
 		VirtualChainId:  c.VirtualChainId,
 		BlockHeight:     blockHeight,
 	})
+}
+
+// provides functionality required for CreateDeployTransaction
+func ReadSourcesFromDir(dirname string) ([][]byte, error) {
+	files, err := ioutil.ReadDir(dirname)
+	if err != nil {
+		return nil, err
+	}
+
+	var sources [][]byte
+	for _, f := range files {
+		if strings.HasSuffix(f.Name(), ".go") && !strings.HasSuffix(f.Name(), "_test.go") {
+			source, err := ioutil.ReadFile(path.Join(dirname, f.Name()))
+			if err != nil {
+				return nil, err
+			}
+
+			sources = append(sources, source)
+		}
+	}
+
+	return sources, nil
 }
